+++ CHANGELOG.md
@@ -1,9 +1,8 @@
-### 0.2.0 (Next)
+### 0.2.0 (7/10/2015)
 
 * Sending `send_message` with nil or empty text will yield `Nothing to see here.` with a GIF instead of `no_text` - [@dblock](https://github.com/dblock).
 * Added support for operators with `operator [name]` - [@dblock](https://github.com/dblock).
 * Added support for custom commands with `command [name]` - [@dblock](https://github.com/dblock).
-* Your contribution here.
 
 ### 0.1.0 (6/2/2015)